Minor Fixes
---
Turns out some of my logic was wrong;
I wrote the code such that if the connection
was active, then the connection should start
rather than continue, it all was due to a little
error in the order of the if statement and
all the fixes of this commit are pretty much the
same style.

Set PLACEHOLDER_NUMBER as 0, changed the
order of the if statement within listen, changed the
order to get the deltas within continueConnection
and added alternative code (commented) to
endConnection.